g2_generateresponse/requirements.txt
flasgger==0.9.7.1
Flask==3.0.3
Flask_Cors==5.0.0
openai==1.43.0
python-dotenv==1.0.1
Requests==2.32.3
requests_toolbelt==1.0.0
